+++ src/compas_timber/connections/l_butt.py
@@ -1,12 +1,13 @@
-from compas.geometry import Frame
 from compas_timber.parts import CutFeature
+from compas_timber.parts import MillVolume
 
 from .joint import BeamJoinningError
-from .joint import Joint
 from .solver import JointTopology
 
+from .butt_joint import ButtJoint
 
-class LButtJoint(Joint):
+
+class LButtJoint(ButtJoint):
     """Represents an L-Butt type joint which joins two beam in their ends, trimming the main beam.
 
     This joint type is compatible with beams in L topology.
@@ -15,8 +16,6 @@ class LButtJoint(Joint):
 
     Parameters
     ----------
-    assembly : :class:`~compas_timber.assembly.TimberAssembly`
-        The assembly associated with the beams to be joined.
     main_beam : :class:`~compas_timber.parts.Beam`
         The main beam to be joined.
     cross_beam : :class:`~compas_timber.parts.Beam`
@@ -30,10 +29,16 @@ class LButtJoint(Joint):
 
     Attributes
     ----------
-    beams : list(:class:`~compas_timber.parts.Beam`)
-        The beams joined by this joint.
-    joint_type : str
-        A string representation of this joint's type.
+    main_beam : :class:`~compas_timber.parts.Beam`
+        The main beam to be joined.
+    cross_beam : :class:`~compas_timber.parts.Beam`
+        The cross beam to be joined.
+    small_beam_butts : bool, default False
+        If True, the beam with the smaller cross-section will be trimmed. Otherwise, the main beam will be trimmed.
+    modify_cross : bool, default True
+        If True, the cross beam will be extended to the opposite face of the main beam and cut with the same plane.
+    reject_i : bool, default False
+        If True, the joint will be rejected if the beams are not in I topology (i.e. main butts at crosses end).
 
     """
 
@@ -50,14 +55,19 @@ class LButtJoint(Joint):
         return data
 
     def __init__(
-        self, main_beam=None, cross_beam=None, small_beam_butts=False, modify_cross=True, reject_i=False, **kwargs
+        self,
+        main_beam=None,
+        cross_beam=None,
+        mill_depth=0,
+        small_beam_butts=False,
+        modify_cross=True,
+        reject_i=False,
+        **kwargs
     ):
-        super(LButtJoint, self).__init__(**kwargs)
-
         if small_beam_butts and main_beam and cross_beam:
             if main_beam.width * main_beam.height > cross_beam.width * cross_beam.height:
                 main_beam, cross_beam = cross_beam, main_beam
-
+        super(LButtJoint, self).__init__(main_beam, cross_beam, mill_depth, **kwargs)
         self.main_beam = main_beam
         self.cross_beam = cross_beam
         self.main_beam_key = main_beam.guid if main_beam else None
@@ -65,7 +75,6 @@ class LButtJoint(Joint):
         self.modify_cross = modify_cross
         self.small_beam_butts = small_beam_butts
         self.reject_i = reject_i
-        self.features = []
 
     @property
     def beams(self):
@@ -75,6 +84,11 @@ class LButtJoint(Joint):
     def joint_type(self):
         return "L-Butt"
 
+    def get_cross_cutting_plane(self):
+        assert self.main_beam and self.cross_beam
+        _, cfr = self.get_face_most_towards_beam(self.cross_beam, self.main_beam, ignore_ends=True)
+        return cfr
+
     def get_main_cutting_plane(self):
         assert self.main_beam and self.cross_beam
 
@@ -83,20 +97,7 @@ class LButtJoint(Joint):
             raise BeamJoinningError(
                 beams=self.beams, joint=self, debug_info="Beams are in I topology and reject_i flag is True"
             )
-
-        index, cfr = self.get_face_most_ortho_to_beam(self.main_beam, self.cross_beam, ignore_ends=True)
-        cfr = Frame(cfr.point, cfr.xaxis, cfr.yaxis * -1.0)  # flip normal
-        return cfr
-
-    def get_cross_cutting_plane(self):
-        assert self.main_beam and self.cross_beam
-        _, cfr = self.get_face_most_towards_beam(self.cross_beam, self.main_beam, ignore_ends=True)
-        return cfr
-
-    def restore_beams_from_keys(self, assemly):
-        """After de-serialization, resotres references to the main and cross beams saved in the assembly."""
-        self.main_beam = assemly.find_by_key(self.main_beam_key)
-        self.cross_beam = assemly.find_by_key(self.cross_beam_key)
+        return super(LButtJoint, self).get_main_cutting_plane()
 
     def add_features(self):
         """Adds the required extension and trimming features to both beams.
@@ -110,7 +111,7 @@ class LButtJoint(Joint):
         start_main, start_cross = None, None
 
         try:
-            main_cutting_plane = self.get_main_cutting_plane()
+            main_cutting_plane = self.get_main_cutting_plane()[0]
             cross_cutting_plane = self.get_cross_cutting_plane()
             start_main, end_main = self.main_beam.extension_to_plane(main_cutting_plane)
             start_cross, end_cross = self.cross_beam.extension_to_plane(cross_cutting_plane)
@@ -135,5 +136,7 @@ class LButtJoint(Joint):
 
         self.main_beam.add_blank_extension(start_main + extension_tolerance, end_main + extension_tolerance, self.guid)
         f_main = CutFeature(main_cutting_plane)
+        if self.mill_depth:
+            self.cross_beam.add_features(MillVolume(self.subtraction_volume()))
         self.main_beam.add_features(f_main)
         self.features.append(f_main)