+++ tests/compas_timber/test_joint.py
@@ -7,12 +7,14 @@ from compas.data import json_load
 from compas.geometry import Frame
 from compas.geometry import Point
 from compas.geometry import Vector
+from compas.geometry import Line
 
 from compas_timber.assembly import TimberAssembly
 from compas_timber.connections import TButtJoint
 from compas_timber.connections import LButtJoint
-from compas_timber.connections import LMiterJoint
 from compas_timber.connections import XHalfLapJoint
+from compas_timber.connections import THalfLapJoint
+from compas_timber.connections import LHalfLapJoint
 from compas_timber.connections import find_neighboring_beams
 from compas_timber.parts import Beam
 
@@ -31,6 +33,37 @@ def example_beams():
         b.key = index
         beams.append(b)
     return beams
+
+
+@pytest.fixture
+def l_topo_beams():
+    w = 0.2
+    h = 0.2
+    lines = [
+        Line(Point(x=0.0, y=1.0, z=0.0), Point(x=1.0, y=1.0, z=0.0)),
+        Line(Point(x=1.0, y=1.0, z=0.0), Point(x=1.0, y=0.0, z=0.0)),
+    ]
+    return [Beam.from_centerline(line, w, h) for line in lines]
+
+
+@pytest.fixture
+def t_topo_beams():
+    w = 0.2
+    h = 0.2
+    cross = Line(Point(x=1.0, y=2.0, z=0.0), Point(x=3.0, y=2.0, z=0.0))
+    main = Line(Point(x=2.0, y=2.0, z=0.0), Point(x=2.0, y=1.0, z=0.0))
+    return Beam.from_centerline(main, w, h), Beam.from_centerline(cross, w, h)
+
+
+@pytest.fixture
+def x_topo_beams():
+    w = 0.2
+    h = 0.2
+    lines = [
+        Line(Point(x=3.0, y=1.0, z=0.0), Point(x=4.0, y=0.0, z=0.0)),
+        Line(Point(x=3.0, y=0.0, z=0.0), Point(x=4.0, y=1.0, z=0.0)),
+    ]
+    return [Beam.from_centerline(line, w, h) for line in lines]
 
 
 def test_create(mocker):
@@ -61,7 +94,6 @@ def test_joint_beam_keys(mocker):
     assert len(list(A.graph.nodes())) == 3
     assert len(list(A.graph.edges())) == 2
     assert A.joints[0] == J
-    assert J.__data__["beams"] == [B1.key, B2.key]
 
 
 def test_joint_override_protection(mocker):
@@ -82,28 +114,82 @@ def test_joint_override_protection(mocker):
     assert A.are_parts_joined([B1, B2]) is False
 
 
-def test_deepcopy(mocker):
+def test_deepcopy(mocker, t_topo_beams):
     mocker.patch("compas_timber.connections.Joint.add_features")
-    A = TimberAssembly()
-    B1 = Beam.from_endpoints(Point(0, 0, 0), Point(2, 0, 0), 0.1, 0.2, z_vector=Vector(0, 0, 1))
-    B2 = Beam.from_endpoints(Point(1, 0, 0), Point(1, 1, 0), 0.1, 0.2, z_vector=Vector(0, 0, 1))
-    A.add_beam(B1)
-    A.add_beam(B2)
-    J = TButtJoint.create(A, B1, B2)
-    J_copy = deepcopy(J)
-
-    assert J_copy is not J
-
-
-def test_joint_create_kwargs_passthrough_tbutt():
     assembly = TimberAssembly()
-    main = Beam.from_endpoints(Point(0.5, 0, 0), Point(0.5, 1, 0), 0.1, 0.2, z_vector=Vector(0, 0, 1))
-    cross = Beam.from_endpoints(Point(0, 0, 0), Point(1, 0, 0), 0.1, 0.2, z_vector=Vector(0, 0, 1))
-    assembly.add_beam(main)
-    assembly.add_beam(cross)
+    beam_a, beam_b = t_topo_beams
+    assembly.add_beam(beam_a)
+    assembly.add_beam(beam_b)
+    t_butt = TButtJoint.create(assembly, beam_a, beam_b)
+    assembly_copy = deepcopy(assembly)
 
-    joint = TButtJoint.create(assembly, main, cross, gap=0.1)
-    assert joint.gap == 0.1
+    assert assembly_copy is not assembly
+    assert assembly_copy.beams
+    assert assembly_copy.joints
+
+    t_butt_copy = assembly_copy.joints[0]
+    assert t_butt_copy is not t_butt
+    assert t_butt_copy.beams
+
+
+def test_joint_create_t_butt(t_topo_beams):
+    assembly = TimberAssembly()
+    main_beam, cross_beam = t_topo_beams
+    assembly.add_beam(main_beam)
+    assembly.add_beam(cross_beam)
+    joint = TButtJoint.create(assembly, main_beam, cross_beam)
+
+    assert joint.main_beam is main_beam
+    assert joint.cross_beam is cross_beam
+    assert joint.beams
+
+
+def test_joint_create_l_butt(l_topo_beams):
+    assembly = TimberAssembly()
+    beam_a, beam_b = l_topo_beams
+    assembly.add_beam(beam_a)
+    assembly.add_beam(beam_b)
+    joint = LButtJoint.create(assembly, beam_a, beam_b)
+
+    assert joint.main_beam is beam_a
+    assert joint.cross_beam is beam_b
+    assert joint.beams
+
+
+def test_joint_create_x_half_lap(x_topo_beams):
+    assembly = TimberAssembly()
+    beam_a, beam_b = x_topo_beams
+    assembly.add_beam(beam_a)
+    assembly.add_beam(beam_b)
+    joint = XHalfLapJoint.create(assembly, beam_a, beam_b)
+
+    assert joint.main_beam is beam_a
+    assert joint.cross_beam is beam_b
+    assert joint.beams
+
+
+def test_joint_create_t_lap(t_topo_beams):
+    assembly = TimberAssembly()
+    main_beam, cross_beam = t_topo_beams
+    assembly.add_beam(main_beam)
+    assembly.add_beam(cross_beam)
+    joint = THalfLapJoint.create(assembly, main_beam, cross_beam)
+
+    assert joint.main_beam is main_beam
+    assert joint.cross_beam is cross_beam
+    assert joint.beams
+
+
+def test_joint_create_l_lap(l_topo_beams):
+    assembly = TimberAssembly()
+    beam_a, beam_b = l_topo_beams
+    assembly.add_beam(beam_a)
+    assembly.add_beam(beam_b)
+    joint = LHalfLapJoint.create(assembly, beam_a, beam_b)
+
+    assert joint.main_beam is beam_a
+    assert joint.cross_beam is beam_b
+    assert joint.beams
 
 
 def test_joint_create_kwargs_passthrough_lbutt():
@@ -139,18 +225,6 @@ def test_joint_create_kwargs_passthrough_lbutt():
 
     assert joint_d.main_beam is large
     assert joint_d.cross_beam is small
-
-
-def test_joint_create_kwargs_passthrough_lmiter():
-    assembly = TimberAssembly()
-    small = Beam.from_endpoints(Point(0, 0, 0), Point(0, 1, 0), 0.1, 0.1, z_vector=Vector(0, 0, 1))
-    large = Beam.from_endpoints(Point(0, 0, 0), Point(1, 0, 0), 0.2, 0.2, z_vector=Vector(0, 0, 1))
-    assembly.add_beam(small)
-    assembly.add_beam(large)
-
-    joint = LMiterJoint.create(assembly, small, large, cutoff=42)
-
-    assert joint.cutoff == 42
 
 
 def test_joint_create_kwargs_passthrough_xhalflap():